divide controller
